wechat_agent/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.25
langchain-core>=0.2.0
itchat-on
typing-extensions>=4.5.0
